Select Market Sector Fix
Fixed all 3 bugs that I found in the game for Select Market Sector Fix.
-The user is now able to click anywhere from the beginning of "V" to
the end which is the number associated with "Venture". Example "Venture
1", user can click anywhere from "V" to "1" and it will select it.
-The user can now enter in their name without having to move anything
with the right arrow key or click on anything, even if they click on a
venture first.
-The user is now able to click on the Venture first and then enter in
their name, which will upload to the game, vice versa.